fix: moved test/page.tsx to root page.tsx
