+++ src/app/page.tsx
@@ -1,21 +1,29 @@
-import { Header } from "@/components/header";
-import { Hero } from "@/components/hero";
-import { SocialProof } from "@/components/social-proof";
-import { Pricing } from "@/components/pricing";
-import { Testimonials } from "@/components/testimonials";
-import { CtaSection } from "@/components/cta-section";
-import { Footer } from "@/components/footer";
+import { Header2 } from "@/components/header-2";
+import { Hero2 } from "@/components/hero-2";
+import { SocialProof2 } from "@/components/social-proof-2";
+import { Features } from "@/components/features-section";
+import { Component } from "@/components/component";
+import { Component2 } from "@/components/component-2";
+import { Component3 } from "@/components/component-3";
+import { Component4 } from "@/components/component-4";
+import { CtaSection2 } from "@/components/cta-section-2";
+import { CtaSection3 } from "@/components/cta-section-3";
+import { Footer2 } from "@/components/footer-2";
 
-export default function HomePage() {
+export default function TestPage() {
   return (
     <>
-      <Header />
-      <Hero />
-      <SocialProof />
-      <Pricing />
-      <Testimonials />
-      <CtaSection />
-      <Footer />
+      <Header2 />
+      <Hero2 />
+      <SocialProof2 />
+      <Features />
+      <Component />
+      <Component2 />
+      <Component3 />
+      <Component4 />
+      <CtaSection2 />
+      <CtaSection3 />
+      <Footer2 />
     </>
   );
 }
